refact : AlbumService 및 Controller 리펙토링
코드 최적화 및 null 단순화

+++ src/main/java/com/bitharmony/comma/domain/album/file/service/FileService.java
@@ -68,36 +68,36 @@ public class FileService {
 			.build();
 	}
 
-	public void updateFile(MultipartFile multipartFile, String filePath, String bucketName) {
-		ObjectMetadata objectMetadata = new ObjectMetadata();
-		objectMetadata.setContentLength(multipartFile.getSize());
-		objectMetadata.setContentType(multipartFile.getContentType());
-
-		try (InputStream inputStream = multipartFile.getInputStream()) {
-			String keyName = getFileName(filePath, bucketName);
-
-			// S3에 폴더 및 파일 업데이트
-			/**
-			 * Amazon S3에서는 동일한 이름(키)을 가진 파일을 업로드하면 기존 파일을 덮어씁니다.
-			 * 따라서, 같은 이름의 새로운 파일을 업로드하면 기존 파일이 자동으로 업데이트됩니다.
-			 * 별도로 삭제 과정을 거칠 필요는 없습니다.
-			 */
-			amazonS3Client.putObject(
-				new PutObjectRequest(bucketName, keyName, inputStream, objectMetadata).withCannedAcl(
-					CannedAccessControlList.PublicRead));
-
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-	}
+	// public void updateFile(MultipartFile multipartFile, String filePath, String bucketName) {
+	// 	ObjectMetadata objectMetadata = new ObjectMetadata();
+	// 	objectMetadata.setContentLength(multipartFile.getSize());
+	// 	objectMetadata.setContentType(multipartFile.getContentType());
+	//
+	// 	try (InputStream inputStream = multipartFile.getInputStream()) {
+	// 		String keyName = getFileName(filePath, bucketName);
+	//
+	// 		// S3에 폴더 및 파일 업데이트
+	// 		/**
+	// 		 * Amazon S3에서는 동일한 이름(키)을 가진 파일을 업로드하면 기존 파일을 덮어씁니다.
+	// 		 * 따라서, 같은 이름의 새로운 파일을 업로드하면 기존 파일이 자동으로 업데이트됩니다.
+	// 		 * 별도로 삭제 과정을 거칠 필요는 없습니다.
+	// 		 */
+	// 		amazonS3Client.putObject(
+	// 			new PutObjectRequest(bucketName, keyName, inputStream, objectMetadata).withCannedAcl(
+	// 				CannedAccessControlList.PublicRead));
+	//
+	// 	} catch (IOException e) {
+	// 		e.printStackTrace();
+	// 	}
+	// }
 
 	public void deleteFile(String filePath, String bucketName) {
-		System.out.println("deleted" + filePath);
+		if(filePath == null) return;
 		amazonS3Client.deleteObject(new DeleteObjectRequest(bucketName, getFileName(filePath, bucketName)));
 	}
 
 	public Optional<MultipartFile> checkFileByType(MultipartFile multipartFile, FileType fileType) {
-		if (multipartFile.getContentType().startsWith(fileType.getType())) {
+		if (multipartFile != null && multipartFile.getContentType().startsWith(fileType.getType())) {
 			return Optional.of(multipartFile);
 		}
 		return Optional.empty();
